Network-domain basic modeling finished.

+++ src/main/kotlin/com/etsija/ourav2/domain/model/Sleep.kt
@@ -22,7 +22,7 @@ data class Sleep(
     val hrAverage: Double,
 
     @JsonProperty("hr_lowest")
-    val hrLowest: Int,
+    val hrLowest: Double,
 
     @JsonProperty("score")
     val score: Int
